[reverse] replace SYCL buffers with USM for more accurate kernel timing

+++ reverse-hip/main.cu
@@ -48,7 +48,7 @@ int main(int argc, char* argv[]) {
   // bound the number of reverse operations
   std::uniform_int_distribution<int> distribution(100, 9999);
 
-  double time = 0.0;
+  long time = 0;
 
   for (int i = 0; i < iteration; i++) {
     const int count = distribution(generator);
